+++ src/SubscriptionBuilder.php
@@ -41,7 +41,7 @@ final class SubscriptionBuilder
     protected bool $async = false;
 
     public function __construct(
-        #[\Readonly] protected EventManager $eventManager
+        protected readonly EventManager $eventManager,
     ) {}
 
     /**